Enable importing apply-setters Processor
Move ApplySettersProcessor into applysetters module so that it can be
imported.
